сделан туториал + восстановление очков крутости исправлен баг со спавном

+++ Model/EntityModel/Enemy.cs
@@ -4,7 +4,7 @@ namespace Game.Model.EntityModel
 {
     public class Enemy : Entity
     {
-        public Enemy(int x, int y, Action<Entity> onEnemyOnEntityDied, int health, int damage) : base(x, y, onEnemyOnEntityDied, health, damage)
+        public Enemy(int x, int y, Action<Entity> onEnemyOnEntityDied, double health, double damage) : base(x, y, onEnemyOnEntityDied, health, damage)
         {
         }
